fix: move tag tests to separate file to avoid pg-cache conflicts
- Create dedicated tag-functionality.test.ts to isolate tag tests
- Remove tag tests from forked-deployment-scenarios.test.ts
- Fixes 'Cannot add to cache after it has been closed' CI errors
